Add XCode project file for tsclient.

+++ app/tsclient/Parameters.cpp
@@ -43,8 +43,8 @@ void Parameters::parse_options(int argc, char* argv[])
         exit(EXIT_SUCCESS);
     }
 
-    int input_sources = vm.count("shm-identifier") + vm.count("input-archive") +
-                        vm.count("subscribe");
+    size_t input_sources = vm.count("shm-identifier") +
+                           vm.count("input-archive") + vm.count("subscribe");
     if (input_sources == 0)
         throw ParametersException("no input source specified");
     if (input_sources > 1)
